Test correct formatting in HttpClient query
